pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muse-cli"
version = "1.0.0"
description = "A philosophical art search tool - interpret abstract text into art search keywords using AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "muse-cli contributors"}
]
keywords = ["art", "search", "AI", "philosophy", "gemini", "apify", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-genai>=0.2.0",
    "apify-client>=1.6.0",
    "urllib3>=2.0.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/ewijaya/muse-cli"
Repository = "https://github.com/ewijaya/muse-cli"
Issues = "https://github.com/ewijaya/muse-cli/issues"

[project.scripts]
muse = "main:app"

[tool.setuptools]
py-modules = ["main", "interpreter", "curator", "gallery_apis", "usage_tracker"]
